implement `must_let!` to bind values using syntax mimicking `let` bindings
I wanted to use unsafe_get as a case study to learn more about Rust
macros. I implemented a new macro, `must_let!`, which binds variables
based on patterns instead of producing a value.

The advantages are that it is easier (for me at least) to read
`must_let!` compared to `get!` due to the resemblance to native Rust
syntax, and `must_let!` can be used to access multiple values.

There are some irritating limitations in the implementation here which
are noted in the documentation with `must_let!`.  It might be possible
to overcome these with more knowledge of macros than I currently posses,
or by using procedural macros.

While working on this I found a crate with similar functionality,
[enum_extract](https://github.com/mystor/enum_extract), which provides
a similar pattern syntax to `must_let!`. A key difference is that
enum_extract requires the user to specify code for the `match` branch
that does not match the given pattern.